Refactored code to have separate CLI and logic
- rhino-agnostic cli builder and logic module
- added build script for rhino
- multiple targets
