+++ builder/pystubsbuilder/SharedAssemblyInfo.cs
@@ -1,0 +1,10 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyCompany("Robert McNeel & Associates")]
+[assembly: AssemblyCopyright("Copyright © 2019")]
+
+[assembly: ComVisible(false)]
+
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
